pyproject.toml
[tool.poetry]
name = "cocktailnet"
version = "0.1.0"
description = ""
authors = ["Ivan Valles Perez <ivallesp@amazon.com>"]

[tool.poetry.dependencies]
python = "3.7.3"
pandas = "^1.1.4"
Pillow = "^8.0.1"
matplotlib = "^3.3.3"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
